Module turbopack_core::chunk

Modules

availability_info

available_assets

optimize

Traits and functions to optimize a list of chunks.

Structs

ChunkContentResult

ChunkData

ChunkDataOption

ChunkDataOptionVc

Vc for ChunkDataOption

ChunkDataVc

Vc for ChunkData

ChunkGroupReference

A reference to multiple chunks from a [ChunkGroup]

ChunkGroupReferenceVc

Vc for ChunkGroupReference

ChunkItemVc

ChunkItems

ChunkItemsVc

Vc for ChunkItems

ChunkVc

ChunkableModuleReferenceVc

ChunkableModuleVc

ChunkingContextVc

ChunkingTypeOption

ChunkingTypeOptionVc

Vc for ChunkingTypeOption

Chunks

ChunksData

ChunksDataVc

Vc for ChunksData

ChunksVc

Vc for Chunks

EvaluatableAssetVc

EvaluatableAssets

EvaluatableAssetsVc

Vc for EvaluatableAssets

ModuleIdVc

Vc for ModuleId

ModuleIds

A list of module ids.

ModuleIdsVc

Vc for ModuleIds

OutputChunkRuntimeInfo

Aggregated information about a chunk content that can be used by the runtime code to optimize chunk loading.

OutputChunkRuntimeInfoVc

Vc for OutputChunkRuntimeInfo

OutputChunkVc

PassthroughAssetVc

Enums

ChunkingType

Specifies how a chunk interacts with other chunks when building a chunk group

ModuleId

A module id, which can be a number or string

Traits

Chunk

A chunk is one type of asset. It usually contains multiple chunk items.

ChunkItem

ChunkableModule

An Asset that can be converted into a Chunk.

ChunkableModuleReference

An AssetReference implementing this trait and returning true for [ChunkableModuleReference::is_chunkable] are considered as potentially chunkable references. When all Assets of such a reference implement ChunkableModule they are placed in Chunks during chunking. They are even potentially placed in the same Chunk when a chunk type specific interface is implemented.

ChunkingContext

A context for the chunking that influences the way chunks are created

EvaluatableAsset

Marker trait for the chunking context to accept evaluated entries.

FromChunkableModule

OutputChunk

PassthroughAsset

An Asset that should never be placed into a chunk, but whose references should still be followed.

Functions

chunk_content

chunk_content_split

Type Definitions

ChunkGroupReferenceReadRef

see [turbo_tasks::ReadRef]

ChunkItemTraitRef

see [turbo_tasks::TraitRef]

ChunkItemsReadRef

see [turbo_tasks::ReadRef]

ChunkTraitRef

see [turbo_tasks::TraitRef]

ChunkableModuleReferenceTraitRef

see [turbo_tasks::TraitRef]

ChunkableModuleTraitRef

see [turbo_tasks::TraitRef]

ChunkingTypeOptionReadRef

see [turbo_tasks::ReadRef]

ChunksReadRef

see [turbo_tasks::ReadRef]

ModuleIdReadRef

see [turbo_tasks::ReadRef]

ModuleIdsReadRef

see [turbo_tasks::ReadRef]

OutputChunkRuntimeInfoReadRef

see [turbo_tasks::ReadRef]

OutputChunkTraitRef

see [turbo_tasks::TraitRef]