Trait turbopack_core::chunk::ChunkableAssetReference

pub trait ChunkableAssetReference: AssetReference + ValueToString {
    fn chunking_type(&self, _context: ChunkingContextVc) -> ChunkingTypeOptionVc;
}

An AssetReference implementing this trait and returning true for [ChunkableAssetReference::is_chunkable] are considered as potentially chunkable references. When all Assets of such a reference implement ChunkableAsset they are placed in Chunks during chunking. They are even potentially placed in the same Chunk when a chunk type specific interface is implemented.

Required Methods

fn chunking_type(&self, _context: ChunkingContextVc) -> ChunkingTypeOptionVc