Trait turbopack_core::chunk::EvaluatableAsset

pub trait EvaluatableAsset: Asset + Module + ChunkableModulewhere
    TaskInput: for<'a> From<&'a Self>,{ }

Marker trait for the chunking context to accept evaluated entries.

The chunking context implementation will resolve the dynamic entry to a well-known value or trait object.