Module turbopack_core::reference

Re-exports

pub use source_map::SourceMapReferenceVc;

Modules

source_map

Structs

AssetReferenceVc

AssetReferences

Multiple AssetReferences

AssetReferencesVc

Vc for AssetReferences

SingleAssetReference

A reference that always resolves to a single asset.

SingleAssetReferenceVc

Vc for SingleAssetReference

Traits

AssetReference

A reference to one or multiple Assets or other special things. There are a bunch of optional traits that can influence how these references are handled. e. g. ChunkableAssetReference or ParallelChunkReference

Functions

all_assets

Aggregates all Assets referenced by an Asset including transitively referenced Assets. This basically gives all Assets in a subgraph starting from the passed Asset.

all_assets_inline

Aggregates all Assets referenced by an Asset including transitively referenced Assets. This basically gives all Assets in a subgraph starting from the passed Asset.

all_referenced_assets

Aggregates all Assets referenced by an Asset. AssetReference This does not include transitively references Assets, but it includes primary and secondary Assets referenced.

all_referenced_assets_inline

Aggregates all Assets referenced by an Asset. AssetReference This does not include transitively references Assets, but it includes primary and secondary Assets referenced.

primary_referenced_assets

Aggregates all primary Assets referenced by an Asset. AssetReference This does not include transitively references Assets, only includes primary Assets referenced.

primary_referenced_assets_inline

Aggregates all primary Assets referenced by an Asset. AssetReference This does not include transitively references Assets, only includes primary Assets referenced.

Type Definitions

AssetReferenceTraitRef

see turbo_tasks::TraitRef

AssetReferencesReadRef

see turbo_tasks::ReadRef

SingleAssetReferenceReadRef

see turbo_tasks::ReadRef