Trait turbopack_core::reference::AssetReference

pub trait AssetReference: ValueToString {
    fn resolve_reference(&self) -> ResolveResultVc;
}

A reference to one or multiple Assets or other special things. There are a bunch of optional traits that can influence how these references are handled. e. g. ChunkableAssetReference or ParallelChunkReference

Required Methods

fn resolve_reference(&self) -> ResolveResultVc