Module turbopack_core::reference_type

Structs

CommonJsReferenceSubTypeVc

Vc for CommonJsReferenceSubType

CssReferenceSubTypeVc

Vc for CssReferenceSubType

EcmaScriptModulesReferenceSubTypeVc

Vc for EcmaScriptModulesReferenceSubType

EntryReferenceSubTypeVc

Vc for EntryReferenceSubType

InnerAssets

Named references to inner assets. Modules can used them to allow to per-module aliases of some requests to already created module assets. Name is usually in UPPER_CASE to make it clear that this is an inner asset.

InnerAssetsVc

Vc for InnerAssets

ReferenceTypeVc

Vc for ReferenceType

TypeScriptReferenceSubTypeVc

Vc for TypeScriptReferenceSubType

UrlReferenceSubTypeVc

Vc for UrlReferenceSubType

Enums

CommonJsReferenceSubType

CssReferenceSubType

EcmaScriptModulesReferenceSubType

EntryReferenceSubType

ReferenceType

TypeScriptReferenceSubType

UrlReferenceSubType

Type Definitions

CommonJsReferenceSubTypeReadRef

see [turbo_tasks::ReadRef]

CssReferenceSubTypeReadRef

see [turbo_tasks::ReadRef]

EcmaScriptModulesReferenceSubTypeReadRef

see [turbo_tasks::ReadRef]

EntryReferenceSubTypeReadRef

see [turbo_tasks::ReadRef]

InnerAssetsReadRef

see [turbo_tasks::ReadRef]

ReferenceTypeReadRef

see [turbo_tasks::ReadRef]

TypeScriptReferenceSubTypeReadRef

see [turbo_tasks::ReadRef]

UrlReferenceSubTypeReadRef

see [turbo_tasks::ReadRef]