Module turbopack_core::resolve

Modules

options

origin

parse

pattern

Structs

AffectingResolvingAssetReference

AffectingResolvingAssetReferenceVc

Vc for AffectingResolvingAssetReference

AliasMap

A map of AliasPatterns to the [Template]s they resolve to.

AliasMapIntoIter

An owning iterator over the entries of an AliasMap.

AliasMapLookupIterator

An iterator over the aliases that match a request.

FindContextFileResultVc

Vc for FindContextFileResult

ResolveAliasMap

Content of an “alias” configuration

ResolveAliasMapVc

Vc for ResolveAliasMap

ResolveResultVc

Vc for ResolveResult

Enums

AliasMatch

Result of a lookup in the alias map.

AliasPattern

An alias pattern.

ExportsValue

The result an “exports” field describes. Can represent multiple alternatives, conditional result, ignored result (null mapping) and a plain result.

FindContextFileResult

ResolveResult

SpecialType

Traits

AliasTemplate

A trait for types that can be used as a template for an alias.

Functions

find_context_file

find_context_file_inline

handle_resolve_error

package_json

package_json_inline

resolve

resolve_inline

resolve_raw

resolve_raw_inline

Type Definitions

AffectingResolvingAssetReferenceReadRef

see [turbo_tasks::ReadRef]

FindContextFileResultReadRef

see [turbo_tasks::ReadRef]

ResolveResultReadRef

see [turbo_tasks::ReadRef]