Module turbopack_core::resolve::pattern

Structs

PatternMatches

PatternMatchesVc

Vc for PatternMatches

PatternVc

Vc for Pattern

QueryMap

QueryMapVc

Vc for QueryMap

Enums

Pattern

PatternMatch

Functions

read_matches

Find all files or directories that match the provided pattern with the specified context directory. prefix is the already matched part of the pattern that leads to the context directory. When force_in_context is set, leaving the context directory by matching .. is not allowed.

read_matches_inline

Find all files or directories that match the provided pattern with the specified context directory. prefix is the already matched part of the pattern that leads to the context directory. When force_in_context is set, leaving the context directory by matching .. is not allowed.

Type Definitions

PatternMatchesReadRef

see [turbo_tasks::ReadRef]

PatternReadRef

see [turbo_tasks::ReadRef]

QueryMapReadRef

see [turbo_tasks::ReadRef]