Function turbopack_core::resolve::pattern::read_matches

pub fn read_matches(
    context: FileSystemPathVc,
    prefix: String,
    force_in_context: bool,
    pattern: PatternVc
) -> PatternMatchesVc

Find all files or directories that match the provided pattern with the specified context directory. prefix is the already matched part of the pattern that leads to the context directory. When force_in_context is set, leaving the context directory by matching .. is not allowed.

Symlinks will not be resolved. It’s expected that the caller resolves symlinks when they are interested in that.