Trait turbopack_core::resolve::plugin::ResolvePlugin

pub trait ResolvePlugin {
    // Required methods
    fn after_resolve_condition(&self) -> ResolvePluginConditionVc;
    fn after_resolve(
        &self,
        fs_path: FileSystemPathVc,
        context: FileSystemPathVc,
        request: RequestVc
    ) -> ResolveResultOptionVc;
}

Required Methods

fn after_resolve_condition(&self) -> ResolvePluginConditionVc

A condition which determines if the hooks gets called.

fn after_resolve( &self, fs_path: FileSystemPathVc, context: FileSystemPathVc, request: RequestVc ) -> ResolveResultOptionVc

This hook gets called when a full filepath has been resolved and the condition matches. If a value is returned it replaces the resolve result.