Struct turbopack_core::resolve::AliasMap

pub struct AliasMap<T> { /* private fields */ }

A map of AliasPatterns to the [Template]s they resolve to.

If a pattern has a wildcard character (*) within it, it will capture any number of characters, including path separators. The result of the capture will then be passed to the template.

If the pattern does not have a wildcard character, it will only match the exact string, and return the template as-is.

Implementations

impl<T> AliasMap<T>

pub fn new() -> Self

Creates a new alias map.

pub fn lookup<'a>(&'a self, request: &'a str) -> AliasMapLookupIterator<'a, T>where
T: Debug,

Looks up a request in the alias map.

Returns an iterator to all the matching aliases.

pub fn insert(&mut self, pattern: AliasPattern, template: T) -> Option<T>

Inserts a new alias into the map.

If the map did not have this alias already, None is returned.

If the map had this alias, the template is updated, and the old template is returned.