Struct turbopack_core::resolve::AliasMapLookupIterator

pub struct AliasMapLookupIterator<'a, T> { /* private fields */ }

An iterator over the aliases that match a request.

The items are returned in the order defined by PATTERN_KEY_COMPARE.