Module turbopack_core::source_map

Structs

CrateMapWrapper

Wraps the CrateMap struct so that it can be cached in a Vc.

GenerateSourceMapVc

OptionSourceMap

OptionSourceMapVc

Vc for OptionSourceMap

OptionToken

OptionTokenVc

Vc for OptionToken

OriginalToken

An OriginalToken represents a region of the generated file that exists in user-authored source file.

OriginalTokenVc

Vc for OriginalToken

RegularSourceMap

A regular source map covers an entire file.

SectionMapping

SectionMappingVc

Vc for SectionMapping

SectionedSourceMap

A sectioned source map contains many (possibly recursive) maps covering different regions of the file.

SourceMapAssetReference

A reference to a [SourceMapAsset], used to inform the dev server/build system of the presence of the source map

SourceMapAssetReferenceVc

Vc for SourceMapAssetReference

SourceMapSection

A section of a larger sectioned source map, which applies at source positions >= the offset (until the next section starts).

SourceMapVc

Vc for SourceMap

SyntheticToken

A SyntheticToken represents a region of the generated file that was created by some build tool.

SyntheticTokenVc

Vc for SyntheticToken

TokenVc

Vc for Token

Enums

SourceMap

The source map spec lists 2 formats, a regular format where a single map covers the entire file, and an “index” sectioned format where multiple maps cover different regions of the file.

Token

A token represents a mapping in a source map. It may either be Synthetic, meaning it was generated by some build tool and doesn’t represent a location in a user-authored source file, or it is Original, meaning it represents a real location in source file.

Traits

GenerateSourceMap

Allows callers to generate source maps.

Type Definitions

GenerateSourceMapTraitRef

see turbo_tasks::TraitRef

OptionSourceMapReadRef

see turbo_tasks::ReadRef

OptionTokenReadRef

see turbo_tasks::ReadRef

OriginalTokenReadRef

see turbo_tasks::ReadRef

SectionMappingReadRef

see turbo_tasks::ReadRef

SourceMapReadRef

see turbo_tasks::ReadRef

SyntheticTokenReadRef

see turbo_tasks::ReadRef

TokenReadRef

see turbo_tasks::ReadRef