Module turbopack_core::version

Structs

FileHashVersion

Version implementation that hashes a file at a given path and returns the hex encoded hash as a version identifier.

FileHashVersionVc

Vc for FileHashVersion

NotFoundVersion

NotFoundVersionVc

Vc for NotFoundVersion

PartialUpdate

A partial update to a versioned object.

TotalUpdate

A total update to a versioned object.

UpdateVc

Vc for Update

VersionVc

VersionedAssetContent

A versioned file content.

VersionedAssetContentVc

Vc for VersionedAssetContent

VersionedContentVc

Enums

Update

Describes an update to a versioned object.

Traits

Version

Describes the current version of an object, and how to update them from an earlier version.

VersionedContent

The content of an [Asset] alongside its version.

Type Definitions

FileHashVersionReadRef

see [turbo_tasks::ReadRef]

NotFoundVersionReadRef

see [turbo_tasks::ReadRef]

UpdateReadRef

see [turbo_tasks::ReadRef]

VersionedAssetContentReadRef

see [turbo_tasks::ReadRef]