Trait turbopack_core::asset::Asset

pub trait Asset {
    fn path(&self) -> FileSystemPathVc;
    fn content(&self) -> AssetContentVc;
    fn references(&self) -> AssetReferencesVc;
    fn versioned_content(&self) -> VersionedContentVc;
}

An asset. It also forms a graph when following Asset::references.

Required Methods

fn path(&self) -> FileSystemPathVc

The path of the Asset. It can potentially be a virtual path. It’s not expected to be something you can read to or write from in general, only some Asset types have these property. It’s more like a name/identifier of the Asset.

fn content(&self) -> AssetContentVc

The content of the Asset.

fn references(&self) -> AssetReferencesVc

Other things (most likely Assets) referenced from this Asset.

fn versioned_content(&self) -> VersionedContentVc

The content of the Asset alongside its version.