Module turbopack_core::chunk

Modules

availability_info

available_assets

optimize

Traits and functions to optimize a list of chunks.

Structs

ChunkContentResult

ChunkGroup

ChunkGroupReference

A reference to multiple chunks from a ChunkGroup

ChunkGroupReferenceVc

Vc for ChunkGroupReference

ChunkGroupVc

Vc for ChunkGroup

ChunkItemVc

ChunkItems

ChunkItemsVc

Vc for ChunkItems

ChunkReference

A reference to a Chunk. Can be loaded in parallel, see Chunk.

ChunkReferenceVc

Vc for ChunkReference

ChunkVc

ChunkableAssetReferenceVc

ChunkableAssetVc

ChunkingContextVc

ChunkingTypeOption

ChunkingTypeOptionVc

Vc for ChunkingTypeOption

Chunks

ChunksVc

Vc for Chunks

EvaluatableAssetVc

EvaluatableAssets

EvaluatableAssetsVc

Vc for EvaluatableAssets

EvaluateChunkingContextVc

ModuleIdVc

Vc for ModuleId

ModuleIds

A list of module ids.

ModuleIdsVc

Vc for ModuleIds

ParallelChunkReferenceVc

Enums

ChunkingType

Specifies how a chunk interacts with other chunks when building a chunk group

ModuleId

A module id, which can be a number or string

Traits

Chunk

A chunk is one type of asset. It usually contains multiple chunk items. There is an optional trait ParallelChunkReference that AssetReferences from a Chunk can implement. If they implement that and ParallelChunkReference::is_loaded_in_parallel returns true, all referenced assets (if they are Chunks) are placed in the same chunk group.

ChunkItem

ChunkableAsset

An Asset that can be converted into a Chunk.

ChunkableAssetReference

An AssetReference implementing this trait and returning true for [ChunkableAssetReference::is_chunkable] are considered as potentially chunkable references. When all Assets of such a reference implement ChunkableAsset they are placed in Chunks during chunking. They are even potentially placed in the same Chunk when a chunk type specific interface is implemented.

ChunkingContext

A context for the chunking that influences the way chunks are created

EvaluatableAsset

Marker trait for the chunking context to accept evaluated entries.

EvaluateChunkingContext

Trait for chunking contexts which can generate evaluated chunks.

FromChunkableAsset

ParallelChunkReference

see Chunk for explanation

Functions

chunk_content

chunk_content_split

Type Definitions

ChunkGroupReadRef

see turbo_tasks::ReadRef

ChunkGroupReferenceReadRef

see turbo_tasks::ReadRef

ChunkItemTraitRef

see turbo_tasks::TraitRef

ChunkItemsReadRef

see turbo_tasks::ReadRef

ChunkReferenceReadRef

see turbo_tasks::ReadRef

ChunkTraitRef

see turbo_tasks::TraitRef

ChunkableAssetReferenceTraitRef

see turbo_tasks::TraitRef

ChunkableAssetTraitRef

see turbo_tasks::TraitRef

ChunkingContextTraitRef

see turbo_tasks::TraitRef

ChunkingTypeOptionReadRef

see turbo_tasks::ReadRef

ChunksReadRef

see turbo_tasks::ReadRef

ModuleIdReadRef

see turbo_tasks::ReadRef

ModuleIdsReadRef

see turbo_tasks::ReadRef

ParallelChunkReferenceTraitRef

see turbo_tasks::TraitRef