Trait turbopack_core::chunk::ChunkingContext

pub trait ChunkingContextwhere
    TaskInput: for<'a> From<&'a Self>,{
    // Required methods
    fn context_path(&self) -> FileSystemPathVc;
    fn output_root(&self) -> FileSystemPathVc;
    fn environment(&self) -> EnvironmentVc;
    fn chunk_path(&self, ident: AssetIdentVc, extension: &str) -> FileSystemPathVc;
    fn reference_chunk_source_maps(&self, chunk: AssetVc) -> BoolVc;
    fn can_be_in_same_chunk(&self, asset_a: AssetVc, asset_b: AssetVc) -> BoolVc;
    fn asset_path(&self, content_hash: &str, extension: &str) -> FileSystemPathVc;
    fn with_layer(&self, layer: &str) -> ChunkingContextVc;
    fn generate_chunk(&self, chunk: ChunkVc) -> AssetVc;

    // Provided methods
    fn is_hot_module_replacement_enabled(&self) -> BoolVc { ... }
    fn layer(&self) -> StringVc { ... }
}

A context for the chunking that influences the way chunks are created

Required Methods

fn context_path(&self) -> FileSystemPathVc

fn output_root(&self) -> FileSystemPathVc

fn environment(&self) -> EnvironmentVc

fn chunk_path(&self, ident: AssetIdentVc, extension: &str) -> FileSystemPathVc

fn reference_chunk_source_maps(&self, chunk: AssetVc) -> BoolVc

Reference Source Map Assets for chunks

fn can_be_in_same_chunk(&self, asset_a: AssetVc, asset_b: AssetVc) -> BoolVc

fn asset_path(&self, content_hash: &str, extension: &str) -> FileSystemPathVc

fn with_layer(&self, layer: &str) -> ChunkingContextVc

fn generate_chunk(&self, chunk: ChunkVc) -> AssetVc

Generates an output chunk asset from an intermediate chunk asset.

Provided Methods

fn is_hot_module_replacement_enabled(&self) -> BoolVc

fn layer(&self) -> StringVc