Trait turbopack_core::issue::Issue

pub trait Issue {
    fn severity(&self) -> IssueSeverityVc;
    fn context(&self) -> FileSystemPathVc;
    fn category(&self) -> StringVc;
    fn title(&self) -> StringVc;
    fn description(&self) -> StringVc;
    fn detail(&self) -> StringVc;
    fn documentation_link(&self) -> StringVc;
    fn source(&self) -> OptionIssueSourceVc;
    fn sub_issues(&self) -> IssuesVc;
}

Required Methods

fn severity(&self) -> IssueSeverityVc

Severity allows the user to filter out unimportant issues, with Bug being the highest priority and Info being the lowest.

fn context(&self) -> FileSystemPathVc

The file path that generated the issue, displayed to the user as message header.

fn category(&self) -> StringVc

A short identifier of the type of error (eg “parse”, “analyze”, or “evaluate”) displayed to the user as part of the message header.

fn title(&self) -> StringVc

The issue title should be descriptive of the issue, but should be a single line. This is displayed to the user directly under the issue header.

fn description(&self) -> StringVc

A more verbose message of the issue, appropriate for providing multiline information of the issue.

fn detail(&self) -> StringVc

Full details of the issue, appropriate for providing debug level information. Only displayed if the user explicitly asks for detailed messages (not to be confused with severity).

fn documentation_link(&self) -> StringVc

A link to relevant documentation of the issue. Only displayed in console if the user explicitly asks for detailed messages.

fn source(&self) -> OptionIssueSourceVc

The source location that caused the issue. Eg, for a parsing error it should point at the offending character. Displayed to the user alongside the title/description.

fn sub_issues(&self) -> IssuesVc