Trait turbopack_core::resolve::origin::ResolveOrigin

pub trait ResolveOrigin {
    fn origin_path(&self) -> FileSystemPathVc;
    fn context(&self) -> AssetContextVc;
}

A location where resolving can occur from. It carries some meta information that are needed for resolving from here.

Required Methods

fn origin_path(&self) -> FileSystemPathVc

The origin path where resolving starts. This is pointing to a file, since that might be needed to infer custom resolving options for that specific file. But usually only the directory is relevant for the real resolving.

fn context(&self) -> AssetContextVc

The AssetContext that carries the configuration for building that subgraph.