Trait turbopack_core::source_map::GenerateSourceMap

pub trait GenerateSourceMapwhere
    TaskInput: for<'a> From<&'a Self>,{
    // Required method
    fn generate_source_map(&self) -> OptionSourceMapVc;

    // Provided method
    fn by_section(&self, _section: &str) -> OptionSourceMapVc { ... }
}

Allows callers to generate source maps.

Required Methods

fn generate_source_map(&self) -> OptionSourceMapVc

Generates a usable source map, capable of both tracing and stringifying.

Provided Methods

fn by_section(&self, _section: &str) -> OptionSourceMapVc

Returns an individual section of the larger source map, if found.