Module turbopack_core::version

Structs

FileHashVersion

Version implementation that hashes a file at a given path and returns the hex encoded hash as a version identifier.

FileHashVersionVc

Vc for FileHashVersion

MergeableVersionedContentVc

NotFoundVersion

NotFoundVersionVc

Vc for NotFoundVersion

PartialUpdate

A partial update to a versioned object.

TotalUpdate

A total update to a versioned object.

UpdateVc

Vc for Update

VersionVc

VersionedAssetContent

A versioned file content.

VersionedAssetContentVc

Vc for VersionedAssetContent

VersionedContentMergerVc

VersionedContentVc

VersionedContents

VersionedContentsVc

Vc for VersionedContents

Enums

Update

Describes an update to a versioned object.

Traits

MergeableVersionedContent

This trait allows multiple VersionedContent to declare which VersionedContentMerger implementation should be used for merging.

Version

Describes the current version of an object, and how to update them from an earlier version.

VersionedContent

The content of an [Asset] alongside its version.

VersionedContentMerger

A VersionedContentMerger merges multiple VersionedContent into a single one.

Type Definitions

FileHashVersionReadRef

see turbo_tasks::ReadRef

MergeableVersionedContentTraitRef

see turbo_tasks::TraitRef

NotFoundVersionReadRef

see turbo_tasks::ReadRef

UpdateReadRef

see turbo_tasks::ReadRef

VersionTraitRef

see turbo_tasks::TraitRef

VersionedAssetContentReadRef

see turbo_tasks::ReadRef

VersionedContentMergerTraitRef

see turbo_tasks::TraitRef

VersionedContentTraitRef

see turbo_tasks::TraitRef

VersionedContentsReadRef

see turbo_tasks::ReadRef