Trait turbopack_core::version::VersionedContent

pub trait VersionedContent {
    fn content(&self) -> AssetContentVc;
    fn version(&self) -> VersionVc;
    fn update(&self, from: VersionVc) -> UpdateVc;
}

The content of an [Asset] alongside its version.

Required Methods

fn content(&self) -> AssetContentVc

The content of the [Asset].

fn version(&self) -> VersionVc

Get a unique identifier of the version as a string. There is no way to convert a version identifier back to the original VersionedContent, so the original object needs to be stored somewhere.

fn update(&self, from: VersionVc) -> UpdateVc

Describes how to update the content from an earlier version to the latest available one.